backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.13
pydantic==2.5.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
google-auth==2.23.4
python-multipart==0.0.6
websockets==12.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
factory-boy==3.3.0